requirements.txt
streamlit
pandas
numpy
scikit-learn
langchain
langchain-huggingface
python-dotenv
huggingface_hub
